requirements.txt
behave==1.2.5
cffi==0.9.2
cryptography==0.8.2
enum34==1.0.4
Flask==0.10.1
gnureadline==6.3.3
ipdb==0.8
ipython==3.1.0
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
mock==1.0.1
nose==1.3.6
parse==1.6.6
parse-type==0.3.4
pyasn1==0.1.7
pycparser==2.12
pyOpenSSL==0.15.1
python-etcd==0.3.3
requests==2.6.0
requests-mock==0.6.0
six==1.9.0
urllib3==1.10.3
Werkzeug==0.10.4
